tests/requirements.txt
pytest>=4.0.1,<5.0
pytest-mock>=1.10.4,<2.0
pytest-aiohttp>=0.3.0,<0.4.0